Remove dependency on -fms-extensions and clean up winadapter (#29)
* WIP remove dependency on fms-extensions

* Various fixes

* Narrowing conversions

* More narrowing

* Couple of fixes

* Init guids

* Fixes for win32 builds

* Replace wrong SDK layers header

* Add winadapter guids for sdk layers

* Comments

* TFDML creator ID

* WSL fixes

* WSL fixes

* Remove extra include opts

* Replace uuidof usage with direct IIDs

* Init GUIDs for Windows builds as well

* Don't include winadapter.h in Windows

* Include dxgi

* Arghh

+++ tensorflow/core/kernels/dml_resource_variable_ops.cc
@@ -30,8 +30,9 @@ class DmlUpdateVariableOp : public DmlKernel {
 
   explicit DmlUpdateVariableOp(DmlKernelConstruction* ctx,
                                const InitHelper* init_helper) {
-    uint32_t tensor_sizes[] = {1, 1, 1,
-                               ctx->GetInputTensorShape(1).num_elements()};
+    uint32_t tensor_sizes[] = {
+        1, 1, 1,
+        static_cast<uint32_t>(ctx->GetInputTensorShape(1).num_elements())};
 
     auto tensor_desc = DmlTensorDesc::Create(ctx->GetInputDataType(1),
                                              tensor_sizes, tensor_sizes);
